@@ -3,26 +3,25 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.9.14",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.53",
-    "@material-ui/styles": "^4.9.14",
-    "axios": "^0.19.2",
-    "formik": "^2.1.4",
-    "immer": "^6.0.5",
-    "polished": "^3.5.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-icons": "^3.10.0",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.1",
-    "redux": "^4.0.5",
+    "@material-ui/core": "^4.12.3",
+    "@material-ui/icons": "^4.11.2",
+    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@material-ui/styles": "^4.11.4",
+    "axios": "^0.25.0",
+    "formik": "^2.2.9",
+    "immer": "^9.0.12",
+    "polished": "^4.1.4",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
+    "react-redux": "^7.2.6",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "5.0.0",
+    "redux": "^4.1.2",
     "redux-saga": "^1.1.3",
-    "styled-components": "^5.1.0",
-    "typescript": "~3.7.2",
-    "uuidv4": "^6.0.8",
-    "yup": "^0.28.5"
+    "styled-components": "^5.3.3",
+    "uuid": "^8.3.2",
+    "yup": "^0.32.11"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -33,9 +32,14 @@
     "coverage": "yarn test --coverage --watchAll=false"
   },
   "devDependencies": {
-    "axios-mock-adapter": "^1.18.1",
+    "axios-mock-adapter": "^1.20.0",
     "eslint": "^8.7.0",
-    "prettier": "^2.0.4"
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "redux-devtools-extension": "^2.13.9"
   },
   "eslintConfig": {
     "extends": "react-app"
