@@ -1,8 +1,6 @@
 import styled, { css } from 'styled-components';
-import waterMarkPokeball from 'assets/img/watermark-pokeball.svg';
-import { Text } from 'components/elements';
-
-
+import { Text } from '../../elements';
+import waterMarkPokeball from '../../../assets/img/watermark-pokeball.svg';
 
 export const bgType = {
   grass: css`
@@ -65,9 +63,8 @@ export const Wrapper = styled.div`
   border-radius: 15px;
   padding: 15px;
   height: 130px;
-  ${(props) =>
-    props.loading !== 'loading' &&
-    css`
+  ${(props) => props.loading !== 'loading'
+    && css`
       background: url(${waterMarkPokeball}) right center no-repeat;
     `}
 
