@@ -2,14 +2,12 @@ import React from 'react';
 
 import { Container, ModalStyled } from './styles';
 
-
-
-const Modal = ({ open, children, handleClose }) => {
+function Modal({ open, children, handleClose }) {
   return (
     <ModalStyled open={open} onClose={handleClose}>
       <Container>{children}</Container>
     </ModalStyled>
   );
-};
+}
 
 export default Modal;
