@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import { Pagination } from '@material-ui/lab';
-import { Text } from '../../components/elements';
 import { Grid } from '@material-ui/core';
+import { Text } from '../../components/elements';
 import WatermarkLarge from '../../assets/img/watermark-pokeball-large.svg';
 
 export const Box = styled.div`
