@@ -1,6 +1,4 @@
-
-
-const ApiStatus  = {
+export const ApiStatus = {
   LOADING: 'loading',
   LOADED: 'loaded',
-}
+};
