@@ -1,7 +1,6 @@
-const SharedActionTypes =  {
+export const SharedActionTypes = {
   HANDLE_MODAL: 'pokemon_inf/HANDLE_MODAL',
-}
-
+};
 
 export const Creators = {
   handleModal(value) {
